\name{strrev}
\alias{strrev}
\title{ Reverses a Character String }
\description{
  Returns the reverse of a character string (\code{strrev("abcde")} returns
  \code{"edcba"}). 
}
\usage{
strrev(ss)
}

\arguments{
  \item{ss}{ A string of characters. }
}

\value{
	Returns the reverse of the string.
}
\examples{
strrev("abcde")
}

\keyword{ character }

